+++ pyvkfft/cuda.py
@@ -50,7 +50,9 @@ try:
                                         ctypes.c_int, ctype_int_size_p, ctypes.c_int]
 
     _vkfft_cuda.init_app.restype = ctypes.c_void_p
-    _vkfft_cuda.init_app.argtypes = [_types.vkfft_config, ctypes.POINTER(ctypes.c_int)]
+    _vkfft_cuda.init_app.argtypes = [_types.vkfft_config, ctypes.POINTER(ctypes.c_int),
+                                     ctypes.POINTER(ctypes.c_size_t),
+                                     ctype_int_size_p, ctype_int_size_p]
 
     _vkfft_cuda.fft.restype = ctypes.c_int
     _vkfft_cuda.fft.argtypes = [_types.vkfft_app, ctypes.c_void_p, ctypes.c_void_p]
@@ -83,7 +85,7 @@ class VkFFTApp(VkFFTAppBase):
 
     def __init__(self, shape, dtype: type, ndim=None, inplace=True, stream=None, norm=1,
                  r2c=False, dct=False, dst=False, axes=None, strides=None, tune_config=None,
-                 r2c_odd=False, **kwargs):
+                 r2c_odd=False, verbose=False, **kwargs):
         """
 
         :param shape: the shape of the array to be transformed. The number
@@ -162,6 +164,8 @@ class VkFFTApp(VkFFTAppBase):
                 of 101, the input array should also be padded to 102 (101+1), and
                 the resulting half-Hermitian array also has a size of 51. A
                 flag is thus needed to differentiate the cases of 100+2 or 101+1.
+        :param verbose: if True, print a 1-string info about this VkFFTApp.
+            See __str__ for details.
 
         :raises RuntimeError: if the initialisation fails, e.g. if the CUDA
             driver has not been properly initialised, or if the transform dimensions
@@ -179,9 +183,21 @@ class VkFFTApp(VkFFTAppBase):
         self.config = self._make_config()
         if self.config is None:
             raise RuntimeError("Error creating VkFFTConfiguration. Was the CUDA context properly initialised ?")
+
         res = ctypes.c_int(0)
-        self.app = _vkfft_cuda.init_app(self.config, ctypes.byref(res))
+        # Size of tmp buffer allocated by VkFFT - if any
+        tmp_buffer_nbytes = ctypes.c_size_t(0)
+        # 0 or 1 for each axis, only if the Bluestein algorithm is used
+        use_bluestein_fft = np.zeros(vkfft_max_fft_dimensions(), dtype=int)
+        # number of axis upload per dimension
+        num_axis_upload = np.zeros(vkfft_max_fft_dimensions(), dtype=int)
+
+        self.app = _vkfft_cuda.init_app(self.config, ctypes.byref(res),
+                                        ctypes.byref(tmp_buffer_nbytes),
+                                        use_bluestein_fft, num_axis_upload)
+
         check_vkfft_result(res, shape, dtype, ndim, inplace, norm, r2c, dct, dst, axes, "cuda")
+
         if self.app is None:
             raise RuntimeError("Error creating VkFFTApplication. Was the CUDA driver initialised ?")
         if has_pycuda:
@@ -190,6 +206,14 @@ class VkFFTApp(VkFFTAppBase):
             #  anymore. Except that we cannot be sure this is the right context, if a stream
             #  has been given because we don't have access to cuStreamGetCtx from python...
             self._ctx = cu_drv.Context.get_current()
+
+        self.tmp_buffer_nbytes = np.int64(tmp_buffer_nbytes)
+        self.use_bluestein_fft = [bool(n) for n in use_bluestein_fft[:len(self.shape)]]
+        self.nb_axis_upload = [int(num_axis_upload[i] * (self.skip_axis[i] is False))
+                               for i in range(len(self.shape))]
+
+        if verbose:
+            print(self)
 
     def __del__(self):
         """ Takes care of deleting allocated memory in the underlying