+++ src/vkfft_opencl.cpp
@@ -30,7 +30,8 @@ LIBRARY_API VkFFTConfiguration* make_config(const long*,
                                             const int, const int, const int, const long*,
                                             const int);
 
-LIBRARY_API VkFFTApplication* init_app(const VkFFTConfiguration*, void*, int*);
+LIBRARY_API VkFFTApplication* init_app(const VkFFTConfiguration*, void*, int*,
+                                       size_t*, long*, long*);
 
 LIBRARY_API int fft(VkFFTApplication* app, void*, void*, void*);
 
@@ -205,7 +206,9 @@ VkFFTConfiguration* make_config(const long* size, const size_t fftdim,
 * \param queue: the cl_command_queue
 * \return: the pointer to the newly created VkFFTApplication
 */
-VkFFTApplication* init_app(const VkFFTConfiguration* config, void *queue, int *res)
+VkFFTApplication* init_app(const VkFFTConfiguration* config, void *queue, int *res,
+                           size_t *tmp_buffer_nbytes, long *use_bluestein_fft,
+                           long *num_axis_upload)
 {
   VkFFTApplication* app = new VkFFTApplication({});
   *res = initializeVkFFT(app, *config);
@@ -215,6 +218,14 @@ VkFFTApplication* init_app(const VkFFTConfiguration* config, void *queue, int *r
     delete app;
     return 0;
   }
+  // Did VkFFT allocate a temporary buffer ?
+  if(app->configuration.allocateTempBuffer)
+    *tmp_buffer_nbytes = app->configuration.tempBufferSize[0];
+   else *tmp_buffer_nbytes = 0;
+
+  for(int i=0; i<VKFFT_MAX_FFT_DIMENSIONS; i++) use_bluestein_fft[i] = app->useBluesteinFFT[i];
+  for(int i=0; i<VKFFT_MAX_FFT_DIMENSIONS; i++) num_axis_upload[i] = app->localFFTPlan->numAxisUploads[i];
+
   return app;
 }